create root App component & add logo
